+++ plugins/node/instrumentation-runtime-node/CHANGELOG.md
@@ -1,5 +1,30 @@
 # Changelog
 
+## [0.10.0](https://github.com/open-telemetry/opentelemetry-js-contrib/compare/instrumentation-runtime-node-v0.9.0...instrumentation-runtime-node-v0.10.0) (2024-11-18)
+
+
+### Features
+
+* **deps:** update deps matching '@opentelemetry/*' ([#2535](https://github.com/open-telemetry/opentelemetry-js-contrib/issues/2535)) ([5223a6c](https://github.com/open-telemetry/opentelemetry-js-contrib/commit/5223a6ca10c5930cf2753271e1e670ae682d6d9c))
+
+## [0.9.0](https://github.com/open-telemetry/opentelemetry-js-contrib/compare/instrumentation-runtime-node-v0.8.0...instrumentation-runtime-node-v0.9.0) (2024-11-07)
+
+
+### ⚠ BREAKING CHANGES
+
+* **instrumentation-runtime-node:** add prom-client-metrics ([#2136](https://github.com/open-telemetry/opentelemetry-js-contrib/issues/2136))
+
+### Features
+
+* **instrumentation-runtime-node:** add prom-client-metrics ([#2136](https://github.com/open-telemetry/opentelemetry-js-contrib/issues/2136)) ([80d0c74](https://github.com/open-telemetry/opentelemetry-js-contrib/commit/80d0c74f8cafdfce681ef39e6cea103dca4e4ccd))
+
+## [0.8.0](https://github.com/open-telemetry/opentelemetry-js-contrib/compare/instrumentation-runtime-node-v0.7.0...instrumentation-runtime-node-v0.8.0) (2024-10-25)
+
+
+### Features
+
+* update "@opentelemetry/*" dependencies to 1.27.0/0.54.0 ([2822511](https://github.com/open-telemetry/opentelemetry-js-contrib/commit/2822511a8acffb875ebd67ff2cf95980a9ddc01e))
+
 ## [0.7.0](https://github.com/open-telemetry/opentelemetry-js-contrib/compare/instrumentation-runtime-node-v0.6.0...instrumentation-runtime-node-v0.7.0) (2024-09-02)